Generators are added/removed during interpretation
Components can now be added to a project by adding component name + version (e.g. Harmonizer-1.0.0) as a tag under a GenDef.

The component will be looked up in the components resource directory and its XSD added to the project XSD.

If the tag gets removed, so will the component XSD.